requirements.txt
pandas
tensorflow
sklearn
scipy
numpy
keras
xgboost
nba_py
lxml
bs4
django-bootstrap3
django-tables2
django-filter
django-chartit
djangorestframework
django-extensions
